backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
anthropic>=0.28.0
python-multipart==0.0.9
pydantic[email]==2.7.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0